--- a/packages/fluentui/projects-test/assets/cra/src/index.tsx
+++ b/packages/fluentui/projects-test/assets/cra/src/index.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
+import App from './App';
+
+const root = document.getElementById('root') as HTMLElement;
+
+ReactDOM.render(<App />, root);