@Hotell 2024-12-30
awesome find and fix @layershifter do we have deltas of execution by any chance (before/after) / wondering if it would be worth it to add that heap threshold to v9 test pipelines as well. ty