@layershifter 2025-01-02
@Hotell sure:

```sh
nx run react-northstar:test --skip-nx-cache
```

- With `workerIdleMemoryLimit` (this PR): 104.748 s / 101.12 s /  102.259 
- Without `workerIdleMemoryLimit`: 67.07 s /   64.685 s / 66.258 s

May be tweaking its `value` to something more reasonable (`2048`?) would speed up.